# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pet Shop API",
        "description": "The Pet Shop API offers the required HTTP request methods and endpoints to meet the demands of a FE team creating a suitable UI",
        "contact": {
            "email": "anthonyjoboy2016@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/v1/admin/user-listing": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Display all non-administrative users",
                "description": "Returns a list of all non-administrative usersinformation.",
                "operationId": "listUsers",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "1 user record(s) found."
                                        },
                                        "users": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": 2
                                                    },
                                                    "uuid": {
                                                        "type": "string",
                                                        "example": "754f68bc-b724-4a4c-91da-970cc5592f8e"
                                                    },
                                                    "first_name": {
                                                        "type": "string",
                                                        "example": "Carolyn"
                                                    },
                                                    "last_name": {
                                                        "type": "string",
                                                        "example": "Mann"
                                                    },
                                                    "is_admin": {
                                                        "type": "number",
                                                        "example": 0
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "john.doe@gmail.com"
                                                    },
                                                    "email_verified_at": {
                                                        "type": "string",
                                                        "example": "2022-08-09T22:50:05.000000Z"
                                                    },
                                                    "avatar": {
                                                        "type": "string",
                                                        "example": "null"
                                                    },
                                                    "phone_number": {
                                                        "type": "string",
                                                        "example": "1-831-953-3966"
                                                    },
                                                    "address": {
                                                        "type": "string",
                                                        "example": "75975 Cleora Meadows\\nWest Davebury, CT 86322"
                                                    },
                                                    "is_marketing": {
                                                        "type": "number",
                                                        "example": 0
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2022-08-09T22:50:05.000000Z"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2022-08-09T22:50:05.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "No products records found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/v1/admin/user-delete/{uuid}": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete user record",
                "description": "Delete user information with existing UUID.",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "User UUID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User was successfully deleted."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred while trying to delete user."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An Administrator user cannot be deleted."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    }
}